move function to util.py
